src/main/java/lv/javaguru/productlist/businesslogic/validataion/ProductValidator.java
package lv.javaguru.productlist.businesslogic.validataion;

import lv.javaguru.productlist.domain.Product;

import java.util.List;

import static java.util.stream.Collectors.toList;

public class ProductValidator {

    private List<ProductValidationRuleInterface> validationRules;

    public ProductValidator(List<ProductValidationRuleInterface> rules) {
        validationRules = rules;
    }

    public ProductValidationResponse validate(Product product) {
        List<String> errorMessages =
                validationRules.stream()
                        .filter(rule -> !rule.isValid(product))
                        .map(rule -> rule.errorMessage())
                        .collect(toList());
        if (errorMessages.isEmpty()) {
            return new ProductValidationResponse(true, null);
        } else {
            return new ProductValidationResponse(false, errorMessages);
        }
    }
}
